+++ examples/django_example/factories/upload_upload.py
@@ -15,6 +15,7 @@ from faker_file.providers.epub_file import EpubFileProvider
 from faker_file.providers.ico_file import IcoFileProvider
 from faker_file.providers.jpeg_file import JpegFileProvider
 from faker_file.providers.mp3_file import Mp3FileProvider
+from faker_file.providers.odp_file import OdpFileProvider
 from faker_file.providers.ods_file import OdsFileProvider
 from faker_file.providers.odt_file import OdtFileProvider
 from faker_file.providers.pdf_file import PdfFileProvider
@@ -22,6 +23,7 @@ from faker_file.providers.png_file import PngFileProvider
 from faker_file.providers.pptx_file import PptxFileProvider
 from faker_file.providers.rtf_file import RtfFileProvider
 from faker_file.providers.svg_file import SvgFileProvider
+from faker_file.providers.tar_file import TarFileProvider
 from faker_file.providers.txt_file import TxtFileProvider
 from faker_file.providers.webp_file import WebpFileProvider
 from faker_file.providers.xlsx_file import XlsxFileProvider
@@ -40,6 +42,7 @@ Faker.add_provider(EpubFileProvider)
 Faker.add_provider(IcoFileProvider)
 Faker.add_provider(JpegFileProvider)
 Faker.add_provider(Mp3FileProvider)
+Faker.add_provider(OdpFileProvider)
 Faker.add_provider(OdsFileProvider)
 Faker.add_provider(OdtFileProvider)
 Faker.add_provider(PdfFileProvider)
@@ -47,6 +50,7 @@ Faker.add_provider(PngFileProvider)
 Faker.add_provider(PptxFileProvider)
 Faker.add_provider(RtfFileProvider)
 Faker.add_provider(SvgFileProvider)
+Faker.add_provider(TarFileProvider)
 Faker.add_provider(TxtFileProvider)
 Faker.add_provider(WebpFileProvider)
 Faker.add_provider(XlsxFileProvider)
@@ -81,6 +85,7 @@ PROVIDER_CHOICES = [
     lambda: IcoFileProvider(None).ico_file(storage=STORAGE),
     lambda: JpegFileProvider(None).jpeg_file(storage=STORAGE),
     lambda: Mp3FileProvider(None).mp3_file(storage=STORAGE),
+    lambda: OdpFileProvider(None).odp_file(storage=STORAGE),
     lambda: OdsFileProvider(None).ods_file(storage=STORAGE),
     lambda: OdtFileProvider(None).odt_file(storage=STORAGE),
     lambda: PdfFileProvider(None).pdf_file(storage=STORAGE),
@@ -88,6 +93,7 @@ PROVIDER_CHOICES = [
     lambda: PptxFileProvider(None).pptx_file(storage=STORAGE),
     lambda: RtfFileProvider(None).rtf_file(storage=STORAGE),
     lambda: SvgFileProvider(None).svg_file(storage=STORAGE),
+    lambda: TarFileProvider(None).tar_file(storage=STORAGE),
     lambda: TxtFileProvider(None).txt_file(storage=STORAGE),
     lambda: XlsxFileProvider(None).xlsx_file(storage=STORAGE),
     lambda: ZipFileProvider(None).zip_file(storage=STORAGE),
@@ -99,7 +105,7 @@ def pick_random_provider(*args, **kwargs):
 
 
 class AbstractUploadFactory(DjangoModelFactory):
-    """Base Upload factory."""
+    """Abstract Upload factory."""
 
     name = Faker("text", max_nb_chars=100)
     description = Faker("text", max_nb_chars=1000)
@@ -123,6 +129,7 @@ class UploadFactory(AbstractUploadFactory):
         ico_file = Trait(file=Faker("ico_file", storage=STORAGE))
         jpeg_file = Trait(file=Faker("jpeg_file", storage=STORAGE))
         mp3_file = Trait(file=Faker("mp3_file", storage=STORAGE))
+        odp_file = Trait(file=Faker("odp_file", storage=STORAGE))
         ods_file = Trait(file=Faker("ods_file", storage=STORAGE))
         odt_file = Trait(file=Faker("odt_file", storage=STORAGE))
         pdf_file = Trait(file=Faker("pdf_file", storage=STORAGE))
@@ -130,6 +137,7 @@ class UploadFactory(AbstractUploadFactory):
         pptx_file = Trait(file=Faker("pptx_file", storage=STORAGE))
         rtf_file = Trait(file=Faker("rtf_file", storage=STORAGE))
         svg_file = Trait(file=Faker("svg_file", storage=STORAGE))
+        tar_file = Trait(file=Faker("tar_file", storage=STORAGE))
         txt_file = Trait(file=Faker("txt_file", storage=STORAGE))
         webp_file = Trait(file=Faker("webp_file", storage=STORAGE))
         xlsx_file = Trait(file=Faker("xlsx_file", storage=STORAGE))