+++ src/faker_file/providers/mp3_file/generators/edge_tts_generator.py
@@ -59,6 +59,5 @@ class EdgeTtsMp3Generator(BaseMp3Generator):
             loop = asyncio.get_event_loop_policy().new_event_loop()
             loop.run_until_complete(_generate())
             # Return result
-            _fake_file.seek(0)
             loop.close()
-            return _fake_file.read()
+            return _fake_file.getvalue()