+++ src/faker_file/providers/odt_file.py
@@ -63,7 +63,7 @@ class OdtFileProvider(BaseProvider, FileMixin):
         content: Optional[str] = None,
         **kwargs,
     ) -> StringValue:
-        """Generate a ODT file with random text.
+        """Generate an ODT file with random text.
 
         :param storage: Storage. Defaults to `FileSystemStorage`.
         :param prefix: File name prefix.
@@ -93,9 +93,8 @@ class OdtFileProvider(BaseProvider, FileMixin):
             document = OpenDocumentText()
             document.text.addElement(P(text=content))
             document.save(_fake_file)
-            _fake_file.seek(0)
 
-            storage.write_bytes(filename, _fake_file.read())
+            storage.write_bytes(filename, _fake_file.getvalue())
 
         # Generic
         file_name = StringValue(storage.relpath(filename))